```text refactor: Update Makefile to use environment variable for CODECOV_TOKEN
The Makefile has been updated to use the environment variable $CODECOV_TOKEN instead of ${{ secrets.CODECOV_TOKEN }} for uploading coverage reports to Codecov. This change improves the security of the token by not exposing it in the Makefile. It also aligns with the recent changes in the CI workflow to distribute test runs and upload coverage reports.